# Improve parsing performance by reducing token cloning