@alamb 2024-12-19
This PR has another conflict now

What I am hoping / planning to do is make the suggested doc improvements above and get it merged in. I also think it would be fine to merge it in and make the doc comment improvements as a follow on PR

I'll try to do this if no one else beats me to it