@alamb 2024-12-24
I merged up to resolve a conflict. While reviewing the code I have some ideas on how to simplify it -- I'll try and make a PR later today. It will be a fun coding exercuse