@alamb 2024-12-24
Some follow on PRs:
- https://github.com/apache/datafusion-sqlparser-rs/pull/1617
- https://github.com/apache/datafusion-sqlparser-rs/pull/1618